ndctl, list: move the --human description to an include
In preparation for inject-error, move the --human option description
to a file that can be included. Reword the description improving
the grammar, and making it more generic.

The example in the list man page for --human was also missing a [verse]
keyword, making the --human part of the example appear in a single line
rather than formatted json - fix that.

Cc: Dan Williams <dan.j.williams@intel.com>
Signed-off-by: Vishal Verma <vishal.l.verma@intel.com>
Signed-off-by: Dan Williams <dan.j.williams@intel.com>