ndctl, bash-completion: updates for --mode options
With the updates to --mode for create-namespace, deprecate the old names
from bash completion. The man page removes mention of the old names, so
remove them from the completion output as well to avoid confusion.

Add the new names into the mode filtering scheme alongside the old names
so that the same filtering behaviour is seen regardless of whether a
user supplies old or new names.

Signed-off-by: Vishal Verma <vishal.l.verma@intel.com>
Signed-off-by: Dan Williams <dan.j.williams@intel.com>