ndctl. test: fix module-taint sanity-check
nfit_test_init() validates that the libnvdimm modules that a test would
use are the properly instrumented external versions. If module signing
is enabled the sanity check will fail because the check expects only the
'O' flag, but unsigned external modules will have the 'O' flag and the
'E' flag set. Relax the constraint to just check for 'O'.

Fixes: 00fc65075c89 ("test: validate nfit_test modules...")
Signed-off-by: Dan Williams <dan.j.williams@intel.com>
Signed-off-by: Vishal Verma <vishal.l.verma@intel.com>