ndctl: complete move to "fsdax" and "devdax"
Add on to the work started by:

commit ebb4fb605e68 ("ndctl, create-namespace: introduce "fsdax" and "devdax" modes")

and change some more user visible places to use "fsdax" and "devdax" modes
instead of "memory" and "dax", respectively.  Having multiple terms for the
same mode is confusing for users.

We will continue to accept "memory" and "dax" as parameters, but all output
and man pages will now use the updated terms.

Note that after the above referenced commit we still printed the old names
in the default 'ndctl list' output for backward compatibility with scripts.
This patch intentionally breaks that backward compatibility in favor of
avoiding confusion and using the new mode names everywhere.

Signed-off-by: Ross Zwisler <ross.zwisler@linux.intel.com>
Reviewed-by: Dan Williams <dan.j.williams@intel.com>
Signed-off-by: Vishal Verma <vishal.l.verma@intel.com>