ndctl, inject-error: add a --saturate option
Now that error injection by default injects fewer bytes per block,
provide an option to allow it to behave in the old way, i.e. injecting
the entire block.

The downside of these reworks is that where as the old way injected the
entire address range in one command, this now issues an ndctl command
per block, which can be slower. However, since this command is only for
testing and debugging purposes, it is expected that the user will
perform targeted, small injections rather than expansive ones at a time,
so the one-DSM-per-block change shouldn't be a noticeable hit. If this
changes in the future, we can consider adding an optimization for the
--saturate path.

Cc: Dan Williams <dan.j.williams@intel.com>
Reviewed-by: Dave Jiang <dave.jiang@intel.com>
Signed-off-by: Vishal Verma <vishal.l.verma@intel.com>