ndctl: Remove dependency on linker garbage collection
A sub-set of the unit tests are optionally included in the build when
"--enable-test" is specified to the configure script. However, the "ndctl
test" command support code should not be included in the build when
"--enable-test" is not specified. The linker was being helpful and
throwing away the error, because cmd_test() is never called, but we
should not have wasted time compiling and throwing away that file in the
first place.

Reported-by: Nick Clifton <nickc@redhat.com>
Link: https://bugzilla.redhat.com/show_bug.cgi?id=1598961
Signed-off-by: Dan Williams <dan.j.williams@intel.com>
Signed-off-by: Vishal Verma <vishal.l.verma@intel.com>