ndctl, bash-completion: add completion for ndctl-monitor
Add bash completions for the 'ndctl monitor' command. This introduces
some new semantics where the completion for --log= can be either a file
path, or one of the canonical options (currently 'standard' or 'syslog'.
We handle this by (ab)using compgen to test whether the entered
string is a substring of one of the canonical strings. If it is, then we
include the array of the canonical strngs as well as regular _filedir
output in our reply. If not, then we rely on _filedir doing its thing.

Cc: QI Fuli <qi.fuli@jp.fujitsu.com>
Signed-off-by: Vishal Verma <vishal.l.verma@intel.com>