ndctl: Use max_available_extent for namespace
The available_size attribute returns all the unused regions, but a
memory namespace has to use contiguous free regions. This patch uses the
max_available_extent attribute, which reports the largest capacity that
can be created when determining what size to allocate.

If the attribute is available, this patch will also report the max
available extent in the region list output.

Signed-off-by: Keith Busch <keith.busch@intel.com>
Tested-by: Masayoshi Mizuma <m.mizuma@jp.fujitsu.com>
Signed-off-by: Vishal Verma <vishal.l.verma@intel.com>