ndctl, documentation: Clarify the dimm id for ndctl list d option
A user may be confused by the description for the "dimm id" argument
for ndctl list -d because the id is expected to be the 'nmem' number,
and not the 'id' field that is reported by ndctl list.

  # ndctl list -d nmem7 | jq -r .id
  cdab-0a-07e0-fefffeff
  # ndctl list -d cdab-0a-07e0-fefffeff

The appropriate id here is X in 'nmemX' (Above example, X is 7).
Let's clarify the description.

Signed-off-by: Masayoshi Mizuma <m.mizuma@jp.fujitsu.com>
Signed-off-by: Vishal Verma <vishal.l.verma@intel.com>