ndctl: Create ndctl udev rules for dirty shutdown
This patch provides an ndctl udev rule and the program it runs. The rule
sends two ndctl commands to allow applications to manage unsafe shutdowns.

The first command acknowledges the shutdown. For nvdimms that support
this, the last shutdown status will remain set to its current status
until it is acknowledged.

The second command retrieves the unclean shutdown count (USC) and saves it
in a known location. Only root can directly access the health's shutdown
count, so we have to stash it somewhere accessible to non-privileged
users. A successful execution of the rule will write USC to the run time
tmpfs location. By default, the location will be set to:

  /run/ndctl/nmem<X>/usc

A distro may change this location using the '--with-tmpfilesdir=[DIR]'.

Reading the file will report the count observed when the dimm was
added.

Signed-off-by: Keith Busch <keith.busch@intel.com>
Signed-off-by: Vishal Verma <vishal.l.verma@intel.com>