ndctl: release v62
This release incorporates functionality up to the 4.19 kernel, and a
number of bug fixes and improvements.

Highlights include addition of the 'ndctl monitor' command to monitor
for SMART health events, use of the new max_available_extent sysfs
attribute for namespace creation, verbosity levels for ndctl-list, a
udev rule for enabling the LSS latch when supported, a bypass route
for making the unsafe shutdown count available for non-privileged users,
improvements to ndctl-inject-smart that include an 'uninject' option
for all fields, and a new unit test, a number of static analysis fixes,
and unit test improvements and fixes.

Commands:
    monitor: new command for monitoring SMART health events
    list: support -v, -vv, -vvv verbosity levels
    inject-smart: add --uninject-* and --uninject-all options
    create-namespace: use max_available_extent for namespace creation
    list: add new fields to -H for alarm_enabled_*
    list: always output JSON arrays when --human is absent

Tests:
    dax.sh: dax-poisonCheck for availability of MAP_SYNC
    dax.sh: fix return code
    device-dax: relax canned timeouts
    monitor: new test
    inject-smart: new test
    max_available_extent_ns: new test

APIs:
    ndctl_cmd_smart_inject_ctrl_temperature
    ndctl_dimm_get_event_flags
    ndctl_dimm_get_flags
    ndctl_dimm_get_health
    ndctl_dimm_is_flag_supported
    ndctl_dimm_smart_inject_supported
    ndctl_region_get_max_available_extent